tactalyze-flask-master/tactalyze/requirements.txt
numpy
pandas
matplotlib
fpdf2
openpyxl
flask
python-dotenv
flask-sqlalchemy
pymysql
werkzeug
flask_simplelogin
gunicorn
gevent
